# I've optimized the PulseID generation speed and added benchmarks.